+++ lib/Posts/LangFilter.php
@@ -61,6 +61,12 @@ class LangFilter {
 		$post_lang_table         = ( new PostTable() )->get_table_name();
 		$translatable_post_types = implode( "','", LangInterface::get_translatable_post_types() );
 
+		/**
+		 *  FIXME: possible problem when making a post type untranslatable, the locale info will
+		 *  still be in the table but it should be ignored.
+		 *
+		 * Should we only show untranslatable posts that have the default language or NULL.
+		 */
 		return "SELECT post_id
 			FROM {$post_lang_table} AS PT
 			WHERE locale = '{$current_lang}'
@@ -138,6 +144,7 @@ class LangFilter {
 	/**
 	 * Returns whether the filtering of posts should happen.
 	 *
+	 * @since 0.4.3 - Add check for `ubb_lang_filter` query_var with a false value to stop the lang filter.
 	 * @since 0.0.1
 	 *
 	 * @param WP_Query $query
@@ -168,6 +175,11 @@ class LangFilter {
 
 		// Don't apply filters on switch_to_blog to blogs without the plugin.
 		if ( ! LangInterface::is_unbabble_active() ) {
+			return false;
+		}
+
+		// Stop the language filter via a query_var.
+		if ( ! ( $query->query_vars['ubb_lang_filter'] ?? true ) ) {
 			return false;
 		}