+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Always include `wp-admin/includes/plugin.php` so we have `is_active_plugin()` in the front-end.
+- Bulk edit not working on non english back-offices and for non `post` post types.
+- No language filters not showing up for non `post` post types.
 
 ## [0.4.4] - 2024-07-15